+++ CHANGES.rst
@@ -1,8 +1,8 @@
 Changes
 =======
 
-Unreleased
-----------
+- Added a transaction-manager ``run`` for running a function as a
+  transaction, retrying as necessary on transient errors.
 
 - Fixed the transaction manager ``attempts`` method. It didn't stop
   repeating when there wasn't an error.